+++ Runtime/VisualScripting/Scripts/Nodes/Unit/CheckOwnershipNode.cs
@@ -1,5 +1,4 @@
-﻿using Virtuademy.SDK.Environments.ClientModels;
-
+﻿
 using Unity.VisualScripting;
 
 namespace Virtuademy.SDK.Environments.VisualScripting