+++ Runtime/VisualScripting/Scripts/Nodes/Unit/CollectEventDataNode.cs
@@ -1,4 +1,4 @@
-using Virtuademy.SDK.Environments.ClientModels;
+using Virtuademy.ScriptingApi;
 
 using Unity.VisualScripting;
 
@@ -32,21 +32,21 @@ namespace Virtuademy.SDK.Environments.VisualScripting
 
         protected override void Definition()
         {
-            CMEvent = ValueInput<CMSession>(nameof(CMEvent), null).NullMeansSelf();
+            CMEvent = ValueInput<SessionView>(nameof(CMEvent), null).NullMeansSelf();
 
-            ID = ValueOutput(nameof(ID), (flow) => flow.GetValue<CMSession>(CMEvent).Id);
+            ID = ValueOutput(nameof(ID), (flow) => flow.GetValue<SessionView>(CMEvent).Id);
 
-            Title = ValueOutput(nameof(Title), (flow) => flow.GetValue<CMSession>(CMEvent).Experience.Title);
+            Title = ValueOutput(nameof(Title), (flow) => flow.GetValue<SessionView>(CMEvent).Experience.Title);
 
-            Description = ValueOutput(nameof(Description), (flow) => flow.GetValue<CMSession>(CMEvent).Experience.Description);
+            Description = ValueOutput(nameof(Description), (flow) => flow.GetValue<SessionView>(CMEvent).Experience.Description);
 
-            StartDateTime = ValueOutput(nameof(StartDateTime), (flow) => flow.GetValue<CMSession>(CMEvent).StartDateTime);
+            StartDateTime = ValueOutput(nameof(StartDateTime), (flow) => flow.GetValue<SessionView>(CMEvent).StartDateTime);
 
-            EndDateTime = ValueOutput(nameof(EndDateTime), (flow) => flow.GetValue<CMSession>(CMEvent).EndDateTime);
+            EndDateTime = ValueOutput(nameof(EndDateTime), (flow) => flow.GetValue<SessionView>(CMEvent).EndDateTime);
 
-            Tags = ValueOutput(nameof(Tags), (flow) => flow.GetValue<CMSession>(CMEvent).Tags);
+            Tags = ValueOutput(nameof(Tags), (flow) => flow.GetValue<SessionView>(CMEvent).Tags);
 
-            IsEventPublic = ValueOutput(nameof(IsEventPublic), (flow) => flow.GetValue<CMSession>(CMEvent).IsPublic);
+            IsEventPublic = ValueOutput(nameof(IsEventPublic), (flow) => flow.GetValue<SessionView>(CMEvent).IsPublic);
         }
     }
 }