+++ Runtime/VisualScripting/Scripts/Nodes/Unit/GetLocalizationData.cs
@@ -1,5 +1,4 @@
-﻿using Virtuademy.SDK.Environments.ClientModels;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 
 using Unity.VisualScripting;